Project 1 Commit
Project 1

Design an embedded computer centered around the ATMega32
microcontroller. For input, use a push button. For output, use an LED.
Write a C program that blinks the LED on/off for as long as the push
button is pressed. Initially, use instruction timing to control the LED
on/off rate (for this step, use the internal 1MHz clock). Then, revise
your timing based on one of the ATMega32 internal timers (for this
step, use the external 8MHz crystal). The blinking rate should be 500ms
on and 500ms off.