Project 2
Design an embedded computer centered around the ATMega32
microcontroller. For input, use a keypad. For output, use an LCD. Write
a C program that implements a digital clock, showing MM/DD/YYYY on the
top row and HH:MM:SS on the bottom row. Provide all the UI needed to
set the date/time on your digital clock.